Use PBKDF2 to hash shared secret
We were previously storing the password in plaintext as a cookie. This was insecure, but fairly low risk because in order to compromise the key, an attacker would need to compromise the user's browser (beyond XSS, as it's an HttpOnly key).

Storing the secret in plaintext additionally had the problem of not properly encoding characters that are not allowed in HTTP cookies.

This change hashes passwords with PBKDF2 and stores the result as base64-encoded bytes.

Fixes #166