Fix a SQL bug in file reader
In #284, @danwilhelm spotted a mistake in the reader.go implementation where the populateBuffer reads not only the current chunk but all subsequent chunks in the file and then discards them. There was no reason for reading the remaining chunks as far as I remember, so I think this was just a simple oversight.

This change changes the SQL comparison from >= to =.