Infer content type when it wasn't specified at upload time
If the client didn't specify the content type for a file at upload time, we can try to infer it based on the filename. This is helpful for files like mp4s or mp3s that modern browsers can play natively instead of relying on the user to download it and find a client player.