Fix cache headers for static files
We accidentally broke caching for static files in #452 because the technique we were using to look up last modification times of files broke when we started embedding them in the binary.

As a workaround, we now use the last modification time based on the first time the static serving function is called.

Resolves #496