make file expiry for guest links configurable when creating guest links (#613)
This PR is trying to implement the feature request in #515 by allowing
to configure file expiration for guest links.

To sum up what has been done:
- refactor existing expiration for guest links/files to differentiate
between file
- add additional drop down to guest link creation screen and table
header in overview
- add database column to keep track of configured expiry (as what ever
the drop down defines e.g. 7 days)
- translate database entry to a time frame for upload of files
- adjusted test minimally
- added database migration script to update existing databases


I hope this suffices and is what you had in mind with "I'd accept a PR
that implemented the functionality that allowed the admin to specify the
expiration time for guest uploads.".