Delete references to guest link before deleting the link
For some reason, the mattn driver doesn't catch this violation of the FOREIGN KEY constraint, but ncruces does. The fix is to just switch up the order so that we delete the references before deleting the underlying data. It's in a transaction, so we shouldn't be able to end up with invalid data.