Show expiration time in server's local time
The server is currently showing the expiration time in UTC time zone. For users in time zones behind UTC, this causes the expiration date to display a day behind.

The server's local time is probably going to be closer to the client's time zone, so this change uses the server's time zone to display expiration time on the file info page.